重構 let束縛間の依存関係抽出を型制約抽出から分離
